chore: Add run.py to start the Flask application and file structure
