+++ app/templates/index.html
@@ -119,7 +119,7 @@
 
     <div class="compare-section text-center" style="width: 50%">
       <h3>Compare Resume with Job Description</h3>
-      <form action="{{ url_for('compare') }}" method="GET">
+      <form action="{{ url_for('app.compare') }}" method="GET">
         <input
           type="text"
           class="form-control"
